requirements.txt
fastapi
uvicorn
sqlalchemy
pandas
psycopg2-binary
jinja2
python-dotenv